Refactor package, webpack config to reflect PR feedback
