+++ package.json
@@ -6,20 +6,19 @@
   "files": [
     "config.js",
     "lib",
-    "dist",
-    "src"
+    "dist"
   ],
   "scripts": {
-    "babel": "babel src -d lib",
+    "build-lib": "babel src -d lib",
     "build-full": "webpack --progress --config webpack.config.js --output-filename solid-client-full.js",
     "build-with-rdflib": "webpack --progress --colors --optimize-minimize --optimize-occurrence-order --optimize-dedupe --config webpack.config.js",
     "build-without-rdflib": "webpack --progress --colors --optimize-minimize --optimize-occurrence-order --optimize-dedupe --config webpack-no-rdflib.config.js",
     "build-qunit-resources": "npm run clean && mkdir -p dist/resources && npm run build-full && browserify -r ./test/resources/profile-minimal.js:test-minimal-profile -o dist/resources/test-minimal-profile.js && browserify -r ./test/resources/profile-private.js:test-minimal-prefs -o dist/resources/test-minimal-prefs.js",
-    "build": "npm run clean && mkdir dist && npm run babel && npm run build-with-rdflib && npm run build-without-rdflib",
+    "build": "npm run clean && mkdir dist && npm run build-lib && npm run build-with-rdflib && npm run build-without-rdflib",
     "clean": "rm -rf dist/",
     "standard": "standard src/*",
     "tape": "tape test/unit/*.js",
-    "test": "npm run standard && npm run babel && npm run tape",
+    "test": "npm run standard && npm run tape",
     "qunit": "npm run build-qunit-resources && open test/integration/index.html",
     "preversion": "npm test",
     "postpublish": "git push --follow-tags",
