+++ webpack.config.js
@@ -18,6 +18,14 @@ module.exports = {
   module: {
     loaders: [
       {
+        test: /\.js$/,
+        exclude: /(node_modules)/,
+        loader: 'babel',
+        query: {
+          presets: ['es2015']
+        }
+      },
+      {
         test: /\.json$/,
         loader: 'json'
       }
